// phoneBook/src/app/interfaces/contactModel.ts
/**This Model is common between backend and frontend
 * This Model contains _id which is generated by the server which is used for 
 * CRUD operations
 */
export interface ContactModel {
    name: string;
    phoneNumber: string[];
    email: string[],
    dateOfBirth: string,
    _id:String

}